+++ examples/.joyride/scripts/z_joylib/calva_api.cljs
@@ -1,26 +1,19 @@
 (ns z-joylib.calva-api
   (:require ["vscode" :as vscode]
+            ["ext://betterthantomorrow.calva$v0" :as calva]
             [joyride.core :as joyride]
             [promesa.core :as p]
             [z-joylib.editor-utils :as editor-utils]))
 
 (def oc (joyride.core/output-channel))
-(def calva (vscode/extensions.getExtension "betterthantomorrow.calva"))
-(def calvaApi (-> calva
-                  .-exports
-                  .-v0
-                  (js->clj :keywordize-keys true)))
-
-(defn text-for-ranges-key [ranges-key]
-  (second ((get-in calvaApi [:ranges ranges-key]))))
 
 (defn evaluate-in-session+ [session-key code]
-  (p/let [result ((get-in [:repl :evaluateCode] calvaApi)
+  (p/let [result (calva/repl.evaluateCode
                   session-key
                   code
                   #js {:stdout #(.append oc %)
                        :stderr #(.append oc (str "Error: " %))})]
-         (.-result result)))
+    (.-result result)))
 
 (defn clj-evaluate+ [code]
   (evaluate-in-session+ "clj" code))
@@ -31,7 +24,7 @@
 (defn evaluate+
   "Evaluates `code` in whatever the current session is."
   [code]
-  (evaluate-in-session+ ((get-in calvaApi [:repl :currentSessionKey])) code))
+  (evaluate-in-session+ (calva/repl.currentSessionKey) code))
 
 (defn evaluate-selection+ []
   (p/let [code (editor-utils/current-selection-text)
@@ -41,10 +34,10 @@
 ;; Utils for REPL-ing Joyride code, when connected to a project REPL.
 
 (defn joyride-eval-current-form+ []
-  (vscode/commands.executeCommand "joyride.runCode" (text-for-ranges-key :currentForm)))
+  (vscode/commands.executeCommand "joyride.runCode" (second (calva/ranges.currentForm))))
 
 (defn joyride-eval-top-level-form+ []
-  (vscode/commands.executeCommand "joyride.runCode" (text-for-ranges-key :currentTopLevelForm)))
+  (vscode/commands.executeCommand "joyride.runCode" (second (calva/ranges.currentTopLevelForm))))
 
 ;; Bind to some nice keyboard shortcuts, e.g. like so:
 ;;  {