+++ src/joyride/sci.cljs
@@ -3,6 +3,7 @@
             ["path" :as path]
             ["vscode" :as vscode]
             [clojure.string :as str]
+            [goog.object :as gobject]
             [joyride.db :as db]
             [joyride.config :as conf]
             [sci.configs.funcool.promesa :as pconfig]
@@ -31,6 +32,27 @@
       {:file ns-path
        :source (str (fs/readFileSync path-to-load))})))
 
+(def ^:private extension-namespace-prefix "ext://")
+
+(defn- extract-extension-name [namespace]
+  (subs namespace (count extension-namespace-prefix)))
+
+(defn- active-extension? [namespace]
+  (when (.startsWith namespace extension-namespace-prefix)
+    (let [[extension-name _module-name] (.split (extract-extension-name namespace) "$")
+          extension (vscode/extensions.getExtension extension-name)]
+      (and extension
+           (.-isActive extension)))))
+
+(defn- extension-module [namespace]
+  (let [[extension-name module-name] (.split (extract-extension-name namespace) "$")
+        extension (vscode/extensions.getExtension extension-name)]
+    (when extension
+      (when-let [exports (.-exports extension)]
+        (if module-name
+          (gobject/getValueByKeys exports (.split module-name "."))
+          exports)))))
+
 (def !ctx
   (volatile!
    (sci/init {:classes {'js goog/global
@@ -41,19 +63,36 @@
                                           'extension-context (sci/copy-var db/extension-context joyride-ns)
                                           'invoked-script (sci/copy-var db/invoked-script joyride-ns)
                                           'output-channel (sci/copy-var db/output-channel joyride-ns)})
-              :load-fn (fn [{:keys [namespace opts]}]
+              :load-fn (fn [{:keys [ns libname opts]}]
                          (cond
-                           (symbol? namespace)
-                           (source-script-by-ns namespace)
-                           (string? namespace) ;; node built-in or npm library
-                           (if (= "vscode" namespace)
+                           (symbol? libname)
+                           (source-script-by-ns libname)
+                           (string? libname) ;; node built-in or npm library
+                           (cond
+                             (= "vscode" libname)
                              (do (sci/add-class! @!ctx 'vscode vscode)
-                                 (sci/add-import! @!ctx (symbol (str @sci/ns)) 'vscode (:as opts))
+                                 (sci/add-import! @!ctx ns 'vscode (:as opts))
                                  {:handled true})
-                             (let [mod (js/require namespace)
-                                   ns-sym (symbol namespace)]
+
+                             (active-extension? libname)
+                             (let [module (extension-module libname)
+                                   munged-ns (symbol (munge libname))
+                                   refer (:refer opts)]
+                               (sci/add-class! @!ctx munged-ns module)
+                               (sci/add-import! @!ctx ns munged-ns (:as opts))
+                               (when refer
+                                 (doseq [sym refer]
+                                   (let [prop (gobject/get module sym)
+                                         sub-sym (symbol (str munged-ns "$" sym))]
+                                     (sci/add-class! @!ctx sub-sym prop)
+                                     (sci/add-import! @!ctx ns sub-sym sym))))
+                               {:handled true})
+
+                             :else
+                             (let [mod (js/require libname)
+                                   ns-sym (symbol libname)]
                                (sci/add-class! @!ctx ns-sym mod)
-                               (sci/add-import! @!ctx (symbol (str @sci/ns)) ns-sym
+                               (sci/add-import! @!ctx ns ns-sym
                                                 (or (:as opts)
                                                     ns-sym))
                                {:handled true}))))})))